Extract bsp testing utils to a helper trait
